requirements.txt
beets
watchdog
requests
python-dotenv

# optional beets plugins
langdetect # lyrics
pylast # lastgenre plugin